requirements.txt
psycopg2
python-dotenv
# postgresql 버전 추가 필요
openai
langchain
langchain_community
langchain_openai
langchain_core
langchain_chroma
langchain_community
langchain_openai
langchain_core
supabase